+++ include/tdp/bb/node_R3.h
@@ -20,18 +20,18 @@ class NodeR3 : public BaseNode {
   virtual ~NodeR3() = default;
   virtual std::vector<NodeR3> Branch() const;
   const Box& GetBox() const { return box_;}
-  const Eigen::Vector3d& GetLbArgument() const {return t_lb_;}
-  void SetLbArgument(const Eigen::Vector3d& t) {t_lb_ = t;}
+  const Eigen::Vector3f& GetLbArgument() const {return t_lb_;}
+  void SetLbArgument(const Eigen::Vector3f& t) {t_lb_ = t;}
   virtual uint32_t GetBranchingFactor(uint32_t i) const { return 8;}
   virtual std::string ToString() const;
   virtual std::string Serialize() const;
   std::string GetSpace() const { return "R3"; }
  protected:
   Box box_;
-  Eigen::Vector3d t_lb_;
-  virtual double GetVolume_() const { return box_.GetVolume();}
+  Eigen::Vector3f t_lb_;
+  virtual float GetVolume_() const { return box_.GetVolume();}
 };
 
-std::list<NodeR3> GenerateNotesThatTessellateR3(const Eigen::Vector3d&
-    min, const Eigen::Vector3d& max, double max_side_len); 
+std::list<NodeR3> GenerateNotesThatTessellateR3(const Eigen::Vector3f&
+    min, const Eigen::Vector3f& max, float max_side_len); 
 }
