+++ include/tdp/bb/s3_tessellation.h
@@ -6,14 +6,14 @@
 #include <iostream>
 #include <Eigen/Dense>
 #include <Eigen/Geometry>
-#include <manifold/S.h>
+#include <tdp/manifold/S.h>
 #include <tdp/utils/combinations.h>
 #include <tdp/bb/tetrahedron.h>
 
 namespace tdp {
 
 std::vector<Tetrahedron4D> TessellateS3();
-std::vector<Tetrahedron4D> TessellateS3(const Eigen::Vector4d& north);
+std::vector<Tetrahedron4D> TessellateS3(const Eigen::Vector4f& north);
 void TessellationTest(std::vector<Tetrahedron4D>& tetrahedra, uint32_t Nsamples);
 
 }
