+++ src/bb/numeric_helpers.cpp
@@ -6,13 +6,13 @@
 
 namespace tdp {
 
-double LogSumExp(const Eigen::VectorXd& x) {
-  const double x_max = x.maxCoeff();
+float LogSumExp(const Eigen::VectorXf& x) {
+  const float x_max = x.maxCoeff();
   return log((x.array() - x_max).exp().sum()) + x_max;
 };
 
-double SumExp(const Eigen::VectorXd& x) {
-  const double x_max = x.maxCoeff();
+float SumExp(const Eigen::VectorXf& x) {
+  const float x_max = x.maxCoeff();
   return (x.array() - x_max).exp().sum() * exp(x_max);
 };
